% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budgetary_exp_by_gdp_range.R
\name{get_budgetary_exp_by_gdp_range}
\alias{get_budgetary_exp_by_gdp_range}
\title{Get government budget expenses for municipalities on a rank of gdp.}
\usage{
get_budgetary_exp_by_gdp_range(
  gdp_range = c(1, 100),
  gdp_year = 2018,
  exp_year = 2019,
  account = NULL,
  expense_stage = NULL
)
}
\arguments{
\item{gdp_range}{vector with two positions indicating the gdp rank range}

\item{gdp_year}{vector indicating the year of reference for gdp rank}

\item{exp_year}{vector indicating the reference year}

\item{account}{vector with accounts}

\item{expense_stage}{vector with expense stages c("Despesas Empenhadas", "Despesas Liquidadas","Despesas Pagas")}
}
\value{
A tibble with data about expanses on functions governments
}
\description{
This function retrieves values of budget expenses for municipalities on a rank of gdp..
}
\examples{
get_budgetary_exp_by_gdp_range(gdp_range=c(1,100), gdp_year=2018,exp_year = 2019, account = "Pessoal e Encargos Sociais" )
}
