% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budgetary_expenses_municipality.R
\name{get_budgetary_expenses_municipality}
\alias{get_budgetary_expenses_municipality}
\title{Get budgetary expenses for municipalities}
\usage{
get_budgetary_expenses_municipality(
  year = 2019,
  municipality = NULL,
  account = NULL,
  expense_stage = NULL
)
}
\arguments{
\item{year}{vector indicating the year or reference}

\item{municipality}{vector with state abbreviation}

\item{account}{vector with account names}

\item{expense_stage}{vctor with expense stages c("Despesas Empenhadas", "Despesas Liquidadas","Despesas Pagas")}
}
\value{
A tibble with data about budgetary expenses
}
\description{
This function retrieves values of govenrment functions for given municipalities.
}
\examples{
get_budgetary_expenses_municipality(year = 2019,municipality = 3550308, account = "Pessoal e Encargos Sociais", expense_stage= "Despesas Empenhadas")
}
