% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budgetary_expenses_municipality_state.R
\name{get_budgetary_expenses_municipality_state}
\alias{get_budgetary_expenses_municipality_state}
\title{Get government budget expenses for municipalities}
\usage{
get_budgetary_expenses_municipality_state(
  year = 2019,
  state = NULL,
  account = NULL,
  expense_stage = NULL
)
}
\arguments{
\item{year}{vector indicating the year or reference}

\item{state}{vector with state abbreviation}

\item{account}{vector with account names}

\item{expense_stage}{vctor with expense stages c("Despesas Empenhadas", "Despesas Liquidadas","Despesas Pagas")}
}
\value{
A tibble with data about budgetary expenses
}
\description{
This function retrieves values of budget expenses for given municipalities.
}
\examples{
get_budgetary_expenses_municipality_state(year = 2019,state = "AP", account = "Pessoal e Encargos Sociais", expense_stage="Despesas Pagas")
}
