% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budgetary_rev_by_gdp_range.R
\name{get_budgetary_rev_by_gdp_range}
\alias{get_budgetary_rev_by_gdp_range}
\title{Get government budget revenues for municipalities on a rank of gdp.}
\usage{
get_budgetary_rev_by_gdp_range(
  gdp_range = c(1, 100),
  gdp_year = 2018,
  rev_year = 2019,
  account = NULL
)
}
\arguments{
\item{gdp_range}{vector with two positions indicating the gdp rank range}

\item{gdp_year}{vector indicating the year of reference for gdp rank}

\item{rev_year}{vector indicating the reference year}

\item{account}{vector with accounts}
}
\value{
A tibble with data about budgetary revenues
}
\description{
This function retrieves values of budget expenses for municipalities on a rank of gdp..
}
\examples{
get_budgetary_rev_by_gdp_range(gdp_range=c(1,100), gdp_year=2018, rev_year = 2019, account = "Impostos" )
}
