% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budgetary_rev_by_pop_range.R
\name{get_budgetary_rev_by_pop_range}
\alias{get_budgetary_rev_by_pop_range}
\title{Get government budget revenues for municipalities by population rank}
\usage{
get_budgetary_rev_by_pop_range(
  pop_range = c(1, 100),
  pop_year = 2018,
  rev_year = 2019,
  account = NULL
)
}
\arguments{
\item{pop_range}{vector with two positions indicating the gdp rank range}

\item{pop_year}{vector indicating the year of reference for gdp rank}

\item{rev_year}{vector indicating the reference year}

\item{account}{vector with accounts}
}
\value{
A tibble with data about budgetary expenses
}
\description{
This function retrieves values of government functions for given municipalities by gdp rank.
}
\examples{
get_budgetary_rev_by_pop_range(pop_range=c(1,100), pop_year=2018, rev_year = 2019, account = "Impostos" )
}
