% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budgetary_revenues_municipality.R
\name{get_budgetary_revenues_municipality}
\alias{get_budgetary_revenues_municipality}
\title{Get budgetary revenues for municipalities}
\usage{
get_budgetary_revenues_municipality(
  year = 2019,
  municipality = NULL,
  account = NULL
)
}
\arguments{
\item{year}{vector indicating the reference years}

\item{municipality}{vector with IBGE ids for municipalities}

\item{account}{vector with account names}
}
\value{
A tibble with data about budgetary revenues
}
\description{
This function retrieves values of government functions for given municipalities.
}
