% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distinct_function.R
\name{get_distinct_function}
\alias{get_distinct_function}
\title{Get distinct functions}
\usage{
get_distinct_function(year = 2019, scope = "a")
}
\arguments{
\item{year}{number indicating the year or reference}

\item{scope}{char ("a", "f", "s") indicates the scope "a" - functions and sub- fuctions, "s" just subfunctions, "f" just functions}
}
\value{
A vector with functions and/or subfunctions
}
\description{
This function retrieves a set of government functions and/or sub-functions.
Notice that the functions are distinct.
Users may get all the functions and sub-functions, just the subset of functions
as well as just the subset of sub-functions
}
