% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distinct_rev_account.R
\name{get_distinct_rev_account}
\alias{get_distinct_rev_account}
\title{Get distinct budgetary revenue accounts}
\usage{
get_distinct_rev_account(year = 2019)
}
\arguments{
\item{year}{number indicating the year or reference}
}
\value{
A vector with expense accounts
}
\description{
This function retrieves a set of budgetary expense accounts.
}
\examples{
get_distinct_rev_account()
}
