% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_exp_by_pop_range.R
\name{get_function_exp_by_pop_range}
\alias{get_function_exp_by_pop_range}
\title{Get government functions for municipalities by population rank}
\usage{
get_function_exp_by_pop_range(
  pop_range = c(1, 100),
  pop_year = 2018,
  exp_year = 2019,
  gov_function = NULL,
  expense_stage = NULL
)
}
\arguments{
\item{pop_range}{vector with two positions indicating the gdp rank range}

\item{pop_year}{vector indicating the year of reference for gdp rank}

\item{exp_year}{vector indicating the reference year}

\item{gov_function}{vector with government functions}

\item{expense_stage}{vctor with expense stages c("Despesas Empenhadas", "Despesas Liquidadas","Despesas Pagas")}
}
\value{
A tibble with data about budgetary expenses
}
\description{
This function retrieves values of government functions for given municipalities by gdp rank.
}
\examples{
get_function_exp_by_pop_range(pop_range=c(1,100), pop_year=2018, exp_year = 2019, gov_function = "saúde", expense_stage="Despesas Liquidadas"  )
}
