% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_expenses_municipality.R
\name{get_function_expenses_municipality}
\alias{get_function_expenses_municipality}
\title{Get government functions expenses for municipalities}
\usage{
get_function_expenses_municipality(
  year = 2019,
  municipality = NULL,
  gov_function = NULL,
  expense_stage = NULL
)
}
\arguments{
\item{year}{vector indicating the reference year}

\item{municipality}{vector with ibge municipality id}

\item{gov_function}{vector with government funcions ids}

\item{expense_stage}{vector with expense stages c("Despesas Empenhadas", "Despesas Liquidadas","Despesas Pagas")}
}
\value{
A tibble with data about expanses on functions governments
}
\description{
This function retrieves values of govenrment functions for given municipalities.
}
\examples{
get_function_expenses_municipality(year=2019, municipality = "3550308", gov_function = "saúde" , expense_stage="Despesas Pagas")
}
