% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_expenses_municipality_state.R
\name{get_function_expenses_municipality_state}
\alias{get_function_expenses_municipality_state}
\title{Get government functions for municipalities of a given state vector}
\usage{
get_function_expenses_municipality_state(
  year = 2019,
  state = NULL,
  gov_function = NULL,
  expense_stage = NULL
)
}
\arguments{
\item{year}{vector indicating the reference year}

\item{state}{vector with state abbreviation}

\item{gov_function}{vector with government funcions ids}

\item{expense_stage}{vctor with expense stages c("Despesas Empenhadas", "Despesas Liquidadas","Despesas Pagas")}
}
\value{
A tibble with data about expanses on functions governments
}
\description{
This function retrieves values of govenrment functions for given municipalities.
}
\examples{
get_function_expenses_municipality_state(year=2019, state = "AP", gov_function = "Saúde", expense_stage= "Despesas Pagas")
}
