% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_budgetary_rev_municipality.R
\name{get_perc_budgetary_rev_municipality}
\alias{get_perc_budgetary_rev_municipality}
\title{Get government budgetary revenues percentage for municipalities}
\usage{
get_perc_budgetary_rev_municipality(
  year = 2019,
  municipality = NULL,
  account = NULL
)
}
\arguments{
\item{year}{vector indicating the reference year}

\item{municipality}{vector with ibge municipality id}

\item{account}{vector with accounts}
}
\value{
A tibble with data about budgetary expenses
}
\description{
This function retrieves values of government functions in % for given municipalities.
}
\examples{
get_perc_budgetary_rev_municipality(year=2019, municipality = "3550308", account = "Impostos" )
}
