% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_billing_id.R, R/setup_siconfi.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Define your Project Id}
\format{
An object of class \code{environment} of length 0.

An object of class \code{environment} of length 0.
}
\usage{
pkg.env

pkg.env
}
\arguments{
\item{project_id}{a single character value containing the string. Vectors with longer lengths and non-vectors will trigger an error.}
}
\value{
No return

No return
}
\description{
Define your project billing ids here so all your queries are authenticated and return data, not errors. If using in production or leaving code available at public repositories, dotenv is highly recommended.

Define your project billing ids here so all your queries are authenticated and return data, not errors. If using in production or leaving code available at public repositories, dotenv is highly recommended.
}
\examples{
set_billing_id(<your_project_id>)
setup_siconfi(<your_project_id>)
}
\keyword{datasets}
