% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Overrides basedadosdados::read_sql function}
\usage{
read_sql(query, page_size = pkg.env$page_size)
}
\arguments{
\item{query}{a string containing a valid SQL query.}

\item{page_size}{bigrquery internal, how many rows per page should there be. Defaults to 10000, consider increasing if running into performance issues or big queries.}
}
\value{
A tibble containing the query's output.
}
\description{
Define new default value to page_seize
}
\examples{
read_sql(query,page_size = 1e+04)
}
