% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_siconfi.R
\name{setup_siconfi}
\alias{setup_siconfi}
\title{Define your Project Id and dataset parameters}
\usage{
setup_siconfi(
  project_id = NULL,
  database = "basedosdados.br_me_siconfi",
  tabela_despesa = "municipio_despesas_orcamentarias",
  tabela_receita = "municipio_receitas_orcamentarias",
  tabela_funcao = "municipio_despesas_funcao",
  coluna_estagio = "estagio_bd",
  coluna_id_siconfi = "portaria",
  page_size = 10^4
)
}
\arguments{
\item{project_id}{a single character value containing the string. Vectors with longer lengths and non-vectors will trigger an error.}
}
\value{
No return
}
\description{
Define your project billing ids here so all your queries are authenticated and return data, not errors. If using in production or leaving code available at public repositories, dotenv is highly recommended.
}
\examples{
setup_siconfi(<your_project_id>)
}
